requirements.txt
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
peewee==3.15.4
python-telegram-bot-calendar==1.0.5
transliterate==1.10.2